feat(publish): add support for package access
NPM added a new option to publish that allows the user to specify
whether their package is public or private. This is primarily meant for
use with organizations. Now that organizations are a public feature,
this is a more commonly used option.

Further information on the option can be found in the official
documentation: https://docs.npmjs.com/cli/publish