fix(setAuthToken): use publishConfig registry URL (#48)
If set, prefer the value of the `packageConfig.registry` property
over the value of the registry as set in the user's `.npmrc` file.

In one scenario, a package may fetch its dependencies from a virtual
registry that is an overlay of a private registry over the public npm
registry. Yet, that package is configured to publish directly to the
private registry URL. To account for this scenario we need to get the
value of the private registry URL and configure it within the
`.npmrc` file.

BREAKING CHANGE:

`setAuthToken` will not prefer the URL value of the
`publishConfig.registry` property in a package's `package.json` file.